requirements.txt
Flask
cx-Oracle
python-dotenv